Updated the api. Now call XQuery->getIterator() to get a ResultIterator. Dynamic_context_wrapper has changed. There is a SetVariable and SetContextItem for each atomic type. Made it compile on cygwin.

+++ include/zorba/util/bignum/floatimpl.h
@@ -165,7 +165,7 @@ public:
   FloatImpl ceil() const; 
 
   FloatImpl round() const{
-    return round(0);
+    return round((int32_t)0);
   }
 
   FloatImpl round(Integer aPrecision) const;
